* Update the master's JSON validation. * Rename some config options to make better sense. * Switch from convoluted "deathmatch + teams", "deathmatch + teams + ctf" gametype method to just adding new gametypes. * Remove number_of_teams, team count can only ever be 2 anyway, so team counts are automatically set in team gametypes. * Fix string termination problem in user agent string generation; this prevented servers from advertising. * Further limit the magnitude of the random token the master generates during user registration to avoid database overflow. * "player_password" was not being removed from the config prior to advertising on the master, fixed. * Add some gametype synonyms to config processing.

+++ source/sv_main.cpp
@@ -357,10 +357,10 @@ char* SV_GetUserAgent(void)
 {
    std::stringstream user_agent_stream;
 
-   user_agent_stream << "emp-server"    << "/"
-                     << (version / 100) << "."
-                     << (version % 100) << "."
-                     << (subversion)    << "-"
+   user_agent_stream << "emp-server"           << "/"
+                     << (version / 100)        << "."
+                     << (version % 100)        << "."
+                     << ((uint32_t)subversion) << "-"
                      << (cs_protocol_version);
 
    return strdup(user_agent_stream.str().c_str());
